requirements.txt
amqp==2.5.2
asgiref==3.2.3
bleach==3.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
Django>=2.2.8
djangorestframework==3.11.0
dnspython==1.16.0
docutils==0.15.2
eventlet==0.25.1
greenlet==0.4.15
idna==2.8
importlib-metadata==1.3.0
jeepney==0.4.1
keyring==21.0.0
kombu==4.6.7
mock==3.0.5
monotonic==1.5
more-itertools==8.0.2
path==13.1.0
path.py==12.4.0
pkginfo==1.5.0.1
pycparser==2.19
Pygments==2.5.2
pytz==2019.3
PyYAML==5.2
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
SecretStorage==3.1.1
six==1.13.0
sqlparse==0.3.0
tqdm==4.41.0
twine==3.1.1
urllib3==1.25.7
vine==1.3.0
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
